pyproject.toml
[tool.poetry]
name = "comp-dnds"
version = "0.1.0"
description = "Efficiently computing dN/dS according to the NJ86 method"
authors = ["maxibor <maxime_borry@eva.mpg.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
biopython = "^1.81"
numpy = "^1.26.1"
tqdm = "^4.66.1"
scipy = "^1.11.4"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
